Modified walls to try a combination of UnitTypes
Instead of adding a segment at a time, pass in a vector of UnitTypes and BWEB will iterate through combinations to try and place buildings. Terran and Protoss buildings are ling tight, Zerg are not. Terran walls will always be connected together, while Protoss walls can be two separate pieces.